requirements.txt
protobuf==3.20.1
wheel==0.38.0
setuptools==65.5.0
gym==0.21.0
dm-control==1.0.3.post1
dm-env==1.5
dm-tree==0.1.7
einops==0.7.0
hydra-core==1.2.0
matplotlib==3.5.1
lockfile==0.12.2
mujoco_py==2.0.2.5
numpy==1.22.3
omegaconf
pandas==1.4.2
seaborn==0.11.2
tensorflow==2.8.0
tqdm==4.64.0
transformers==4.39.1
datasets
wandb==0.14.0
gym[atari]==0.21.0
autorom[accept-rom-license]
stable_baselines3[extra]==1.5.0
ale-py==0.7.4
procgen
cloudpickle==2.1.0
# fsspec==2022.1.0
git+https://github.com/denisyarats/dmc2gym.git
torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
torchmetrics==1.2.0
h5py==3.6.0
scikit-learn==1.1.3
gymnasium==0.28.1
dacite==1.8.1
# may results in issues with nle: https://github.com/facebookresearch/nle/issues/246
# afterwards add to bashrc --> LD_LIBRARY_PATH="$LD_LIBRARY_PATH:$CONDA_PREFIX/lib"
# minihack==0.1.5
# opencv-python==4.6.0.66